requirements.txt
streamlit
tensorflow-cpu==2.8.0
numpy>=1.19.5
Keras>=2.4.3
opencv-python-headless>=4.5.1.48
streamlit-webrtc>=0.11.0
av==9.1.1
streamlit-lottie==0.0.3
streamlit-option-menu==0.3.2
